refactor: move out /lcp-context folder from /directives folder
