+++ projects/storefrontlib/cms-components/product/product-images/product-images.module.ts
@@ -11,7 +11,7 @@ import { CmsConfig, provideDefaultConfig } from '@spartacus/core';
 import { OutletModule } from '../../../cms-structure/outlet/index';
 import { CarouselModule } from '../../../shared/components/carousel/index';
 import { MediaModule } from '../../../shared/components/media/media.module';
-import { LcpContextDirectiveModule } from '../../../shared/directives/lcp-context/lcp-context-directive.module';
+import { LcpContextDirectiveModule } from '../../../shared/lcp-context/lcp-context-directive.module';
 import { ProductImagesComponent } from './product-images.component';
 
 @NgModule({
