--- projects/storefrontlib/shared/directives/index.ts
@@ -1,7 +1,0 @@
-/*
- * SPDX-FileCopyrightText: 2025 SAP Spartacus team <spartacus-team@sap.com>
- *
- * SPDX-License-Identifier: Apache-2.0
- */
-
-export * from './lcp-context/index';
