+++ projects/storefrontlib/shared/index.ts
@@ -6,7 +6,7 @@
 
 export * from './components/index';
 export * from './config/index';
-export * from './directives/index';
+export * from './lcp-context/index';
 export * from './pipes/index';
 export * from './services/index';
 export * from './utils/index';
